requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.8
dj-rest-auth==6.0.0
Django==5.0.4
django-allauth==0.63.3
django-cors-headers==4.3.1
django-rest-authtoken==2.1.4
django-rest-swagger==2.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.14.0
identify==2.5.36
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
nodeenv==1.9.0
openapi-codec==1.3.2
packaging==24.0
platformdirs==4.2.2
pre-commit==3.7.1
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.26.2
